Creation of Recipe bootstrapper in order to initialize the app with some sample data. This makes use of the default data.sql file to insert some unit of measures and categories into the database on startup.
I then make use of a bootstrapper to construct an example recipe and save it to the database via the RecipeRepository. I then just quickly modified the index page to show a list of all the Recipes built in this way, i.e. currently only one for now but can be expanded at a later date if required.

This makes use of JPA and hibernate to easily save to and retriecve from the DB without any real implementations needing to be written, only interfaces.